# packages/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Qari AI API
  version: "1.0.0"
  description: >
    Companion Qari backend API. Declared riwayah is Hafs 'an 'Asim (see
    CLAUDE.md and ADR-003). Machine feedback is always mediated by the
    confidence-tier policy (ADR-005); this contract never exposes raw model
    scores to clients.
servers:
  - url: https://api.qari.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
  /auth/signup:
    post:
      tags: [auth]
      summary: Create a new guardian account
      operationId: signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [auth]
      summary: Log in with email + password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/guest-upgrade:
    post:
      tags: [auth]
      summary: Upgrade an anonymous guest session to a full guardian account
      description: >
        Preserves guest-created child profiles and their progress, attaching
        them to the newly created guardian account.
      operationId: upgradeGuestAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestUpgradeRequest'
      responses:
        '200':
          description: Guest session upgraded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /content/passages:
    get:
      tags: [content]
      summary: List available passages for the current content version
      operationId: listPassages
      security:
        - bearerAuth: []
      parameters:
        - name: surahNumber
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 114
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: A page of passages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassagePage'
  /content/passages/{passageId}:
    get:
      tags: [content]
      summary: Get a single passage, including reference audio and word timings
      operationId: getPassage
      security:
        - bearerAuth: []
      parameters:
        - name: passageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The passage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassageDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageCursor:
      name: cursor
      in: query
      schema:
        type: string
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflicting resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10
        locale:
          type: string
          enum: [en, ur, ar]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    GuestUpgradeRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10
    AuthSession:
      type: object
      required: [accessToken, refreshToken, userId]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        userId:
          type: string
          format: uuid
...
    PassagePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PassageSummary'
        nextCursor:
          type: string
          nullable: true
    PassageSummary:
      type: object
      required: [id, surahNumber, ayahStart, ayahEnd, riwayah]
      properties:
        id:
          type: string
          format: uuid
        surahNumber:
          type: integer
        ayahStart:
          type: integer
        ayahEnd:
          type: integer
        riwayah:
          type: string
          enum: [hafs_an_asim]
    PassageDetail:
      allOf:
        - $ref: '#/components/schemas/PassageSummary'
        - type: object
          required: [referenceAudioUrl, reciterId, ayahs, translation]
          properties:
            referenceAudioUrl:
              type: string
              format: uri
              nullable: true
              description: >
                Null until a licensed reciter is cleared (docs/licenses/) —
                see docs/STUBS.md. Never points at fabricated "recitation"
                audio; a generated silent placeholder file if present.
            ayahs:
              type: array
              items:
                $ref: '#/components/schemas/PassageAyah'
            translation:
              $ref: '#/components/schemas/TranslationToggle'
            reciterId:
              type: string
              nullable: true
    PassageAyah:
      type: object
      required: [ayahNumber, words]
      properties:
        ayahNumber:
          type: integer
        words:
          type: array
          items:
            $ref: '#/components/schemas/PassageWord'
    PassageWord:
      type: object
      required: [wordIndex, displayText]
      properties:
        wordIndex:
          type: integer
        displayText:
          type: string
          description: Verbatim Uthmani text, diacritics/pause marks intact (docs/content-tokenization.md).
    TranslationToggle:
      type: object
      required: [available]
      description: >
        Translation text is withheld until a translation source's license
        is confirmed to permit this app's use (see docs/STUBS.md — Tanzil's
        hosted translations are non-commercial-only by default).
      properties:
        available:
          type: boolean
        reason:
          type: string
          nullable: true
        language:
          type: string
          enum: [en, ur]
          nullable: true
        sourceName:
          type: string
          nullable: true
        text:
          type: string
          nullable: true
...